requirements.txt
streamlit
pandas
torch
scikit-learn
openai
Pillow
matplotlib